bridge: widen the checkpoint context and give the distance its own byte
Checkpoint ranges were capped at what one baggage byte could hold: the
context packs an assigned-distance index above a countdown, so
bits(max-1)+bits(max-min) <= 8 rejected anything past 1:16. Sparse
checkpoint schedules were unreachable.

The context is now a fixed two bytes (CheckpointContextBytes). Two bytes
covers every range the design allows, since Max is capped at 256. The
width is deliberately NOT varied with the configured range: a width that
changed with the range would be a varint in all but name and would make
byte totals incomparable between ranges.

The assigned distance also moves out of four spare bits in the payload
type byte into its own byte after it, lifting the distinct-distance
limit from 16 to 256. Reconstruction needs the distance to size the
window Bloom BEFORE it can parse the filter, so it cannot be derived
from a reconstruction that has not happened yet. The type and early-leaf
flags stay where they were; fixed-distance payloads are untouched.

Costs, measured over 50k day-1 traces: forward baggage +1 B/span
(+3.82%), checkpoint payload +1 B/checkpoint (~+1.3%). Nothing else
moves -- checkpoint spans, reverse wire bytes, emitter and promotion
counts are bit-identical.

One silent failure this surfaced: validReverseCheckpointSegment read the
depth varint at Payload[1:], so with the distance byte inserted it
parsed the distance as the depth, the depth check failed, and every
reverse segment was rejected with no error. NewReverseHandler now takes
the field width from the forward handler rather than from the caller --
a caller that forgot would have seen exactly that silent rejection.

+++ bridge/bridge.go
@@ -36,10 +36,11 @@ const (
 // StartResult is the data the simulator needs from OnStart to drive its
 // per-trace metric accumulators (matching trace_simulator.py run_traces).
 type StartResult struct {
-	// CheckpointTTL is the packed outgoing baggage byte in randomized mode:
+	// CheckpointTTL is the packed outgoing baggage context in randomized mode:
 	// assigned-distance index in its high bits, remaining TTL in its low bits.
-	// CheckpointRange.RemainingTTL and AssignedDistance decode those fields.
-	CheckpointTTL byte
+	// CheckpointRange.RemainingTTL and AssignedDistance decode those fields. It
+	// occupies CheckpointRange.ContextBytes bytes on the wire, one or two.
+	CheckpointTTL uint16
 	BaggageFound  bool // parent had handler state -> count this as a baggage call
 	BaggageBytes  int  // 3 + len(packed_br) when BaggageFound, else 0
 	EmitBytes     int  // EMIT_PAYLOAD_BYTES set in this OnStart, 0 if none
